backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
deprecation==2.1.0
fastapi==0.135.2
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
packaging==26.0
pillow==12.1.1
pip==25.2
postgrest==2.28.3
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.7.1
pygments==2.20.0
pyiceberg==0.11.1
pyjwt==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
realtime==2.28.3
requests==2.33.1
rich==14.3.3
six==1.17.0
starlette==1.0.0
storage3==2.28.3
strenum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==15.0.1
yarl==1.23.0
zstandard==0.25.0
